+++ MatrixKit/Controllers/MXKRoomViewController.h
@@ -177,6 +177,11 @@
 @property BOOL scrollHistoryToTheBottomOnKeyboardPresentation;
 
 /**
+ YES (default) to show actions button in document preview. NO otherwise.
+ */
+@property BOOL allowActionsInDocumentPreview;
+
+/**
  This object is defined when the displayed room is left. It is added into the bubbles table header.
  This label is used to display the reason why the room has been left.
  */